Add creation to model helper, use it for the application endpoint
